add testing utils & remove node benchmark
