+++ pallets/ddc-clusters/src/lib.rs
@@ -17,6 +17,8 @@
 
 #[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
+#[cfg(any(feature = "runtime-benchmarks", test))]
+pub mod testing_utils;
 
 #[cfg(test)]
 pub(crate) mod mock;
